Added the NotificationHelper
