+++ app/src/main/java/com/example/niroigensuntharam/elec390application/MainActivity.java
@@ -53,51 +53,6 @@ public class MainActivity extends AppCompatActivity{
         getRoomInfoAsync.execute(dateString);
 
         Initialization();
-
-        sendNotification();
-    }
-
-    public void sendNotification() {
-        //Get an instance of NotificationManager//
-
-        Intent notificationIntent = new Intent(this, MainActivity.class);
-        PendingIntent contentIntent = PendingIntent.getActivity(
-                this,
-                0,
-                notificationIntent, PendingIntent.FLAG_CANCEL_CURRENT);
-
-        NotificationCompat.Builder mBuilder = new NotificationCompat.Builder(this)
-                        .setSmallIcon(R.drawable.desktop_icon)
-                        .setContentTitle("Change Room")
-                        .setContentText("Room in this lab soon")
-                        .setContentIntent(contentIntent);
-
-        NotificationCompat.InboxStyle inboxStyle =
-                new NotificationCompat.InboxStyle();
-
-        String[] events = new String[2];
-
-        inboxStyle.setBigContentTitle("Reminder for Room Description");
-
-        events[0] = "Room number: 823";// + currentRoom.getRoomNumber();
-        events[1] = "Course: COEN 390";// + currentRoom.getNextClass();
-
-        for (int i = 0; i < events.length; i++)
-        {
-            inboxStyle.addLine(events[i]);
-        }
-
-        mBuilder.setStyle(inboxStyle);
-
-        // Gets an instance of the NotificationManager service//
-
-        NotificationManager mNotificationManager = (NotificationManager) getSystemService(Context.NOTIFICATION_SERVICE);
-
-        //When you issue multiple notifications about the same type of event, it’s best practice for your app to try to update an existing notification with this new information, rather than immediately creating a new notification. If you want to update this notification at a later date, you need to assign it an ID. You can then use this ID whenever you issue a subsequent notification. If the previous notification is still visible, the system will update this existing notification, rather than create a new one. In this example, the notification’s ID is 001//
-
-        //NotificationManager.notify().
-
-        mNotificationManager.notify(001, mBuilder.build());
     }
 
     // Will refresh to show the rooms the user can currently go to
